# driver-service-node/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RideLink Driver & Vehicle Service API",
    "version": "1.0.0",
    "description": "Manage driver profiles, vehicles, availability, locations, and eligible driver search."
  },
  "servers": [{ "url": "/api/drivers" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "DriverProfileInput": {
        "type": "object",
        "required": ["profile", "vehicleDetails", "serviceArea"],
        "properties": {
          "profile": {
            "type": "object",
            "required": ["firstName", "lastName", "phone", "licenseNumber", "licenseExpiry"],
            "properties": {
              "firstName": { "type": "string" },
              "lastName": { "type": "string" },
              "phone": { "type": "string" },
              "licenseNumber": { "type": "string" },
              "licenseExpiry": { "type": "string", "format": "date-time" }
            }
          },
          "vehicleDetails": {
            "type": "object",
            "required": ["model", "registrationNumber", "vehicleType", "color", "capacity"],
            "properties": {
              "model": { "type": "string" },
              "registrationNumber": { "type": "string" },
              "vehicleType": { "type": "string" },
              "color": { "type": "string" },
              "capacity": { "type": "integer", "minimum": 1 }
            }
          },
          "serviceArea": { "type": "string" }
        }
      },
      "Driver": {
        "allOf": [
          { "$ref": "#/components/schemas/DriverProfileInput" },
          {
            "type": "object",
            "properties": {
              "driverId": { "type": "string" },
              "userId": { "type": "string" },
              "availabilityStatus": { "type": "string", "enum": ["AVAILABLE", "UNAVAILABLE", "IN_RIDE"] },
              "location": {
                "type": "object",
                "properties": {
                  "latitude": { "type": "number", "minimum": -90, "maximum": 90 },
                  "longitude": { "type": "number", "minimum": -180, "maximum": 180 },
                  "updatedAt": { "type": "string", "format": "date-time" }
                }
              },
              "rating": { "type": "number", "minimum": 0, "maximum": 5 }
            }
          }
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "data": { "nullable": true }
        }
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/profile": {
      "put": {
        "summary": "Create or replace the authenticated driver's profile and vehicle",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DriverProfileInput" } } }
        },
        "responses": {
          "201": { "description": "Profile created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Driver" } } } },
          "200": { "description": "Profile updated" },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Missing or invalid token" },
          "403": { "description": "Role is not allowed" }
        }
      }
    },
    "/profile/{driverId}": {
      "get": {
        "summary": "Get a driver profile",
        "parameters": [{ "name": "driverId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Profile retrieved" },
          "401": { "description": "Missing or invalid token" },
          "403": { "description": "Not the profile owner or an admin" },
          "404": { "description": "Profile not found" }
        }
      }
    },
    "/{driverId}/availability": {
      "patch": {
        "summary": "Update driver availability",
        "parameters": [{ "name": "driverId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["availabilityStatus"],
                "properties": { "availabilityStatus": { "type": "string", "enum": ["AVAILABLE", "UNAVAILABLE", "IN_RIDE"] } }
              }
            }
          }
        },
        "responses": { "200": { "description": "Availability updated" }, "400": { "description": "Validation error" }, "401": { "description": "Missing or invalid token" }, "404": { "description": "Profile not found" } }
      }
    },
    "/{driverId}/location": {
      "patch": {
        "summary": "Update simulated current location",
        "parameters": [{ "name": "driverId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["latitude", "longitude"],
                "properties": {
                  "latitude": { "type": "number", "minimum": -90, "maximum": 90 },
                  "longitude": { "type": "number", "minimum": -180, "maximum": 180 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Location updated with timestamp" }, "400": { "description": "Validation error" }, "401": { "description": "Missing or invalid token" }, "404": { "description": "Profile not found" } }
      }
    },
    "/eligible": {
      "get": {
        "summary": "Find available drivers in a service area, optionally within a radius",
        "parameters": [
          { "name": "serviceArea", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "latitude", "in": "query", "required": false, "schema": { "type": "number", "minimum": -90, "maximum": 90 } },
          { "name": "longitude", "in": "query", "required": false, "schema": { "type": "number", "minimum": -180, "maximum": 180 } },
          { "name": "maxDistanceKm", "in": "query", "required": false, "schema": { "type": "number", "minimum": 0, "maximum": 100, "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Eligible drivers retrieved" },
          "400": { "description": "Invalid or incomplete location query" },
          "401": { "description": "Missing or invalid token" },
          "403": { "description": "Role is not allowed" }
        }
      }
    }
  }
}